Refactor: Modularize Advanced Football Match Analysis Feature
This commit refactors the "other_feature_NUEVO.py" module into a dedicated package located at `modules/advanced_analysis/`.

The goals of this refactoring were to:
- Improve code readability and maintainability by separating concerns.
- Enhance modularity by splitting logic into specific files:
    - `app.py`: Main entry point for the feature.
    - `ui.py`: Streamlit UI rendering and layout.
    - `data_fetcher.py`: Handles data fetching via requests and BeautifulSoup.
    - `selenium_manager.py`: Manages Selenium WebDriver interactions.
    - `utils.py`: Common utility functions.
- Preserve original functionality, including Streamlit caching for performance.

Key changes include:
- Creation of the `modules/advanced_analysis/` package.
- Migration of UI, data fetching, Selenium logic, and utilities into their respective Python files within this package.
- Re-application of Streamlit caching decorators (`st.cache_data`, `st.cache_resource`) in the UI layer.
- Removal of the original `modules/other_feature_NUEVO.py` file.
- Update of the root `app.py` to correctly import and use the refactored feature from its new location.
- Basic static analysis (Flake8) and execution checks were performed to ensure the refactored code is functional.
