+++ modules/advanced_analysis/selenium_manager.py
@@ -1,0 +1,187 @@
+import time
+import re
+import sys # For st.error replacement
+from bs4 import BeautifulSoup
+
+from selenium import webdriver
+from selenium.webdriver.chrome.options import Options as ChromeOptions
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.common.exceptions import TimeoutException, WebDriverException, ElementClickInterceptedException # NoSuchElementException removed
+
+from .utils import format_ah_as_decimal_string_of
+from .data_fetcher import BASE_URL_OF
+
+# --- CONFIGURACIÓN GLOBAL DE SELENIUM ---
+SELENIUM_TIMEOUT_SECONDS_OF = 20
+SELENIUM_POLL_FREQUENCY_OF = 0.2
+
+# --- FUNCIONES DE SELENIUM ---
+
+def prepare_h2h_page_for_scraping(driver, match_url_path):
+    """Navigates to the given H2H page URL and waits for a key element."""
+    if not driver:
+        print("Selenium driver not available for page preparation.", file=sys.stderr)
+        return False
+    
+    full_url = f"{BASE_URL_OF}{match_url_path}"
+    try:
+        driver.get(full_url)
+        WebDriverWait(driver, SELENIUM_TIMEOUT_SECONDS_OF, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(
+            EC.presence_of_element_located((By.ID, "table_v1")) # Wait for a common element
+        )
+        time.sleep(0.5) # Small pause for JS rendering
+        return True
+    except TimeoutException:
+        print(f"Timeout waiting for page elements on {full_url}", file=sys.stderr)
+        return False
+    except WebDriverException as e:
+        print(f"WebDriverException during page preparation for {full_url}: {e}", file=sys.stderr)
+        return False
+    except Exception as e:
+        print(f"Unexpected error during page preparation for {full_url}: {e}", file=sys.stderr)
+        return False
+
+def get_selenium_driver_of():
+    options = ChromeOptions(); options.add_argument("--headless"); options.add_argument("--no-sandbox")
+    options.add_argument("--disable-dev-shm-usage"); options.add_argument("--disable-gpu")
+    options.add_argument("user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 Chrome/116.0.0.0 Safari/537.36")
+    options.add_argument('--blink-settings=imagesEnabled=false'); options.add_argument("--window-size=1920,1080")
+    try: 
+        return webdriver.Chrome(options=options)
+    except WebDriverException as e: 
+        print(f"Error inicializando Selenium driver (OF): {e}", file=sys.stderr)
+        return None
+
+def get_h2h_details_for_original_logic_of(driver_instance, key_match_id_for_h2h_url, rival_a_id, rival_b_id, rival_a_name="Rival A", rival_b_name="Rival B"):
+    if not driver_instance: return {"status": "error", "resultado": "N/A (Driver no disponible H2H OF)"}
+    if not key_match_id_for_h2h_url or not rival_a_id or not rival_b_id: return {"status": "error", "resultado": f"N/A (IDs incompletos para H2H {rival_a_name} vs {rival_b_name})"}
+    
+    url_to_visit = f"{BASE_URL_OF}/match/h2h-{key_match_id_for_h2h_url}" 
+    try:
+        driver_instance.get(url_to_visit)
+        WebDriverWait(driver_instance, SELENIUM_TIMEOUT_SECONDS_OF, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(EC.presence_of_element_located((By.ID, "table_v2")))
+        time.sleep(0.7); soup_selenium = BeautifulSoup(driver_instance.page_source, "html.parser")
+    except TimeoutException: return {"status": "error", "resultado": f"N/A (Timeout esperando table_v2 en {url_to_visit})"}
+    except Exception as e: return {"status": "error", "resultado": f"N/A (Error Selenium en {url_to_visit}: {type(e).__name__})"}
+    
+    if not soup_selenium: return {"status": "error", "resultado": f"N/A (Fallo soup Selenium H2H Original OF en {url_to_visit})"}
+    table_to_search_h2h = soup_selenium.find("table", id="table_v2") 
+    if not table_to_search_h2h: return {"status": "error", "resultado": f"N/A (Tabla v2 para H2H no encontrada en {url_to_visit})"}
+
+    for row in table_to_search_h2h.find_all("tr", id=re.compile(r"tr2_\d+")): 
+        links = row.find_all("a", onclick=True); 
+        if len(links) < 2: continue
+        h2h_row_home_id_m = re.search(r"team\((\d+)\)", links[0].get("onclick", "")); h2h_row_away_id_m = re.search(r"team\((\d+)\)", links[1].get("onclick", ""))
+        if not h2h_row_home_id_m or not h2h_row_away_id_m: continue
+        h2h_row_home_id = h2h_row_home_id_m.group(1); h2h_row_away_id = h2h_row_away_id_m.group(1)
+        
+        if {h2h_row_home_id, h2h_row_away_id} == {str(rival_a_id), str(rival_b_id)}:
+            score_span = row.find("span", class_="fscore_2") 
+            if not score_span or not score_span.text or "-" not in score_span.text: continue
+            score_val = score_span.text.strip().split("(")[0].strip(); g_h, g_a = score_val.split("-", 1)
+            tds = row.find_all("td"); handicap_raw_text = "N/A"; HANDICAP_TD_IDX = 11 
+            if len(tds) > HANDICAP_TD_IDX:
+                cell = tds[HANDICAP_TD_IDX]; d_o = cell.get("data-o") 
+                handicap_raw_text = d_o.strip() if d_o and d_o.strip() not in ["", "-"] else (cell.text.strip() if cell.text.strip() not in ["", "-"] else "N/A")
+            
+            # Formatting the handicap line using the utility function
+            handicap_formatted = format_ah_as_decimal_string_of(handicap_raw_text)
+
+            rol_a_in_this_h2h = "H" if h2h_row_home_id == str(rival_a_id) else "A"
+            return {
+                "status": "found", 
+                "goles_home": g_h.strip(), 
+                "goles_away": g_a.strip(), 
+                "handicap": handicap_formatted, # Use formatted handicap
+                "handicap_raw": handicap_raw_text, # Keep raw handicap if needed elsewhere
+                "rol_rival_a": rol_a_in_this_h2h, 
+                "h2h_home_team_name": links[0].text.strip(), 
+                "h2h_away_team_name": links[1].text.strip()
+            }
+    return {"status": "not_found", "resultado": f"H2H directo no encontrado para {rival_a_name} vs {rival_b_name} en historial (table_v2) de la página de ref. ({key_match_id_for_h2h_url})."}
+
+def click_element_robust_of(driver, by, value, timeout=7):
+    try:
+        element = WebDriverWait(driver, timeout, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(EC.presence_of_element_located((by, value)))
+        WebDriverWait(driver, timeout, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(EC.visibility_of(element))
+        driver.execute_script("arguments[0].scrollIntoView({block: 'center', inline: 'center'});", element); time.sleep(0.3)
+        try: 
+            WebDriverWait(driver, 2, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(EC.element_to_be_clickable((by, value))).click()
+        except (ElementClickInterceptedException, TimeoutException): 
+            driver.execute_script("arguments[0].click();", element)
+        return True
+    except Exception: 
+        return False
+
+def extract_last_match_in_league_of(driver, table_css_id_str, main_team_name_in_table, league_id_filter_value, home_or_away_filter_css_selector, is_home_game_filter):
+    try:
+        if league_id_filter_value:
+            league_checkbox_selector = f"input#checkboxleague{table_css_id_str[-1]}[value='{league_id_filter_value}']"
+            click_element_robust_of(driver, By.CSS_SELECTOR, league_checkbox_selector); time.sleep(1.0)
+        
+        click_element_robust_of(driver, By.CSS_SELECTOR, home_or_away_filter_css_selector); time.sleep(1.0)
+        page_source_updated = driver.page_source; soup_updated = BeautifulSoup(page_source_updated, "html.parser")
+        table = soup_updated.find("table", id=table_css_id_str)
+        if not table: return None
+        
+        count_visible_rows = 0
+        for row in table.find_all("tr", id=re.compile(rf"tr{table_css_id_str[-1]}_\d+")):
+            if row.get("style") and "display:none" in row.get("style","").lower(): continue
+            count_visible_rows +=1
+            if count_visible_rows > 10: break 
+            
+            if league_id_filter_value and row.get("name") != str(league_id_filter_value): continue
+            
+            tds = row.find_all("td"); 
+            if len(tds) < 14: continue
+            
+            home_team_row_el = tds[2].find("a"); away_team_row_el = tds[4].find("a")
+            if not home_team_row_el or not away_team_row_el: continue
+            
+            home_team_row_name = home_team_row_el.text.strip(); away_team_row_name = away_team_row_el.text.strip()
+            team_is_home_in_row = main_team_name_in_table.lower() == home_team_row_name.lower()
+            team_is_away_in_row = main_team_name_in_table.lower() == away_team_row_name.lower()
+            
+            if (is_home_game_filter and team_is_home_in_row) or (not is_home_game_filter and team_is_away_in_row):
+                date_span = tds[1].find("span", {"name": "timeData"}); date = date_span.text.strip() if date_span else "N/A"
+                score_class_re = re.compile(r"fscore_"); score_span = tds[3].find("span", class_=score_class_re); score = score_span.text.strip() if score_span else "N/A"
+                handicap_cell = tds[11]; handicap_raw = handicap_cell.get("data-o", handicap_cell.text.strip()) 
+                
+                if not handicap_raw or handicap_raw.strip() == "-": handicap_raw = "N/A"
+                else: handicap_raw = handicap_raw.strip()
+                
+                return {"date": date, "home_team": home_team_row_name, "away_team": away_team_row_name,"score": score, "handicap_line_raw": handicap_raw}
+        return None
+    except Exception: 
+        return None
+
+def get_main_match_odds_selenium_of(driver):
+    odds_info = {"ah_home_cuota": "N/A", "ah_linea_raw": "N/A", "ah_away_cuota": "N/A", "goals_over_cuota": "N/A", "goals_linea_raw": "N/A", "goals_under_cuota": "N/A"}
+    try:
+        live_compare_div = WebDriverWait(driver, SELENIUM_TIMEOUT_SECONDS_OF, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(EC.presence_of_element_located((By.ID, "liveCompareDiv")))
+        bet365_row_selector = "tr#tr_o_1_8[name='earlyOdds']"; bet365_row_selector_alt = "tr#tr_o_1_31[name='earlyOdds']"
+        table_odds = live_compare_div.find_element(By.XPATH, ".//table[contains(@class, 'team-table-other')]")
+        driver.execute_script("arguments[0].scrollIntoView({block: 'center'});", table_odds); time.sleep(0.5)
+        
+        bet365_early_odds_row = None
+        try: 
+            bet365_early_odds_row = WebDriverWait(driver, 5, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(EC.visibility_of_element_located((By.CSS_SELECTOR, bet365_row_selector)))
+        except TimeoutException: 
+            try: 
+                bet365_early_odds_row = WebDriverWait(driver, 3, poll_frequency=SELENIUM_POLL_FREQUENCY_OF).until(EC.visibility_of_element_located((By.CSS_SELECTOR, bet365_row_selector_alt)))
+            except TimeoutException: 
+                return odds_info 
+        
+        tds = bet365_early_odds_row.find_elements(By.TAG_NAME, "td")
+        if len(tds) >= 11:
+            odds_info["ah_home_cuota"] = tds[2].get_attribute("data-o") or tds[2].text.strip() or "N/A"
+            odds_info["ah_linea_raw"] = tds[3].get_attribute("data-o") or tds[3].text.strip() or "N/A" 
+            odds_info["ah_away_cuota"] = tds[4].get_attribute("data-o") or tds[4].text.strip() or "N/A"
+            odds_info["goals_over_cuota"] = tds[8].get_attribute("data-o") or tds[8].text.strip() or "N/A"
+            odds_info["goals_linea_raw"] = tds[9].get_attribute("data-o") or tds[9].text.strip() or "N/A" 
+            odds_info["goals_under_cuota"] = tds[10].get_attribute("data-o") or tds[10].text.strip() or "N/A"
+    except Exception: 
+        pass # Silently ignore errors as per original, or add logging
+    return odds_info
